refactor: split list and listitem into seperate files
